Fix the build on some platforms by using an explicit cast for Univalue type
Summary: It appears that std::vector::size_type is not implicitely convertible to uint64_t or int64_t on on platforms , so cast it explicitely.

Test Plan:
  ./contrib/teamcity/build-configurations.py build-linux32
  ./contrib/teamcity/build-configurations.py build-native-osx

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Differential Revision: https://reviews.bitcoinabc.org/D11244

+++ src/rpc/avalanche.cpp
@@ -614,7 +614,7 @@ static RPCHelpMan decodeavalanchedelegation() {
             result.pushKV("proofid", delegation.getProofId().ToString());
 
             auto levels = delegation.getLevels();
-            result.pushKV("depth", levels.size());
+            result.pushKV("depth", uint64_t(levels.size()));
 
             UniValue levelsArray(UniValue::VARR);
             for (auto &level : levels) {
